+++ src/Repositories/TokenRepository.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Wearesho\Yii\Repositories;
 
+use Wearesho\Yii\Entities\TokenableEntity;
 use Wearesho\Yii\Exceptions\DeliveryLimitReachedException;
 use Wearesho\Yii\Exceptions\InvalidRecipientException;
 use Wearesho\Yii\Exceptions\InvalidTokenException;
@@ -17,34 +18,15 @@ use Wearesho\Yii\Interfaces\TokenRepositoryConfigInterface;
 
 use Horat1us\Yii\Validation;
 use Wearesho\Delivery;
+use Wearesho\Yii\Models\Token;
 
 class TokenRepository implements TokenRepositoryInterface
 {
-    protected TokenRecordInterface $model;
-
-    protected TokenRepositoryConfigInterface $config;
-
-    protected TokenGeneratorInterface $generator;
-
-    protected Delivery\ServiceInterface $deliveryService;
-
-    /**
-     * TokensRepository constructor.
-     * @param TokenRecordInterface|null $model
-     * @param TokenRepositoryConfigInterface $config
-     * @param TokenGeneratorInterface $generator
-     * @param Delivery\ServiceInterface $service
-     */
     public function __construct(
-        TokenRecordInterface           $model,
-        TokenRepositoryConfigInterface $config,
-        TokenGeneratorInterface        $generator,
-        Delivery\ServiceInterface      $service
+        protected TokenRepositoryConfigInterface $config,
+        protected TokenGeneratorInterface        $generator,
+        protected Delivery\ServiceInterface      $deliveryService
     ) {
-        $this->model = $model;
-        $this->generator = $generator;
-        $this->config = $config;
-        $this->deliveryService = $service;
     }
 
     /**
@@ -57,13 +39,11 @@ class TokenRepository implements TokenRepositoryInterface
      */
     public function push(TokenableEntityInterface $entity): TokenInterface
     {
-        $record = $this->pull($entity->getTokenRecipient());
+        $record = $this->pull($entity);
         if (!$record instanceof TokenRecordInterface) {
-            $class = get_class($this->model);
-            /** @var TokenRecordInterface $record */
-            $record = new $class;
-
-            $record->setRecipient($entity->getTokenRecipient());
+            $record = new Token();
+            $record->type = $entity->getTokenType();
+            $record->setRecipient($entity->getRecipient());
             $record->setToken($this->generator->getToken());
         }
 
@@ -94,7 +74,13 @@ class TokenRepository implements TokenRepositoryInterface
             );
         }
 
-        $this->deliveryService->send($token);
+        $entityWithToken = new TokenableEntity(
+            $entity->getRecipient(),
+            str_replace('{token}', $token->getToken(), $entity->getText()),
+            $entity->getTokenType(),
+            $entity->getTokenData()
+        );
+        $this->deliveryService->send($entityWithToken);
 
         if ($token instanceof TokenRecordInterface) {
             $token->increaseDeliveryCount();
@@ -105,12 +91,13 @@ class TokenRepository implements TokenRepositoryInterface
     /**
      * Pulling active token to process it (for example, sending sms)
      */
-    public function pull(string $tokenRecipient): ?TokenRecordInterface
+    public function pull(TokenableEntityInterface $entity): ?TokenRecordInterface
     {
         /** @noinspection PhpIncompatibleReturnTypeInspection */
-        return $this->model->find()
+        return Token::find()
+            ->andWhere(['=', 'token.type', $entity->getTokenType()])
             ->notExpired($this->config->getExpirePeriod())
-            ->whereRecipient($tokenRecipient)
+            ->whereRecipient($entity->getRecipient())
             ->one();
     }
 
@@ -125,12 +112,12 @@ class TokenRepository implements TokenRepositoryInterface
      * @throws InvalidTokenException
      * @throws Validation\Failure
      */
-    public function verify(string $tokenRecipient, string $token): TokenInterface
+    public function verify(TokenableEntityInterface $entity, string $token): TokenInterface
     {
-        $record = $this->pull($tokenRecipient);
+        $record = $this->pull($entity);
 
         if (!$record instanceof TokenRecordInterface) {
-            throw new InvalidRecipientException($tokenRecipient);
+            throw new InvalidRecipientException($entity->getRecipient());
         }
 
         $record->increaseVerifyCount();
@@ -151,5 +138,18 @@ class TokenRepository implements TokenRepositoryInterface
         $record->delete();
 
         return $record;
+    }
+
+    public function delete(TokenableEntityInterface $entity): void
+    {
+        $record = $this->pull($entity);
+
+        if (!$record instanceof TokenRecordInterface) {
+            throw new InvalidRecipientException($entity->getRecipient());
+        }
+
+        if ($record instanceof Token) {
+            $record->delete();
+        }
     }
 }